refactor: 최종 이익 계산식 로직 수정
